Use same address for console I/O
The console is mapped to $FF00 on the simple machine.
Offset addresses CON_IN and CON_OUT define the addresses
used for input and output.

Output was written to $FF00 + CON_OUT = $FF00, whereas input
was read from $FF00 + CON_IN = $FF02. This commit changes
CON_IN to 0 so that $FF00 is now used for both reading and writing.